requirements.txt
torch>=2.4.0,<2.5.0
transformers>=4.36.0,<4.46.0
trl>=0.7.0,<0.12.0
accelerate>=0.25.0
deepspeed>=0.12.0
datasets>=2.14.0
pyyaml
pydantic>=2.0.0
pandas>=2.0.0
pyarrow>=12.0.0
numpy<2
